Integer to Long i some places

+++ minhashing/src/main/java/hashing/HashFunction.java
@@ -4,5 +4,5 @@ package hashing;
  * Created by m on 10.03.15.
  */
 public interface HashFunction {
-	public int hash(int number);
+	public long hash(long number);
 }
